Remove old UpdateIndexHandler and related files
