--- index-handlers/src/test/resources/dynamoDbRecordApplicableEvent.json
@@ -1,106 +1,0 @@
-{
-  "eventID": "e6d0a4122fad36f07da8d61e10ca9f05",
-  "eventName": "INSERT",
-  "eventVersion": "1.1",
-  "eventSource": "aws:dynamodb",
-  "awsRegion": "eu-west-1",
-  "dynamodb": {
-    "approximateCreationDateTime": "2023-08-29T13:07:03.000+00:00",
-    "keys": {
-      "PrimaryKeyRangeKey": {
-        "s": "CANDIDATE#167ee953-b8a9-4ef6-a3c7-a62519ed6a6e"
-      },
-      "PrimaryKeyHashKey": {
-        "s": "CANDIDATE#167ee953-b8a9-4ef6-a3c7-a62519ed6a6e"
-      }
-    },
-    "newImage": {
-      "identifier": {
-        "s": "167ee953-b8a9-4ef6-a3c7-a62519ed6a6e"
-      },
-      "SecondaryIndex1HashKey": {
-        "s": "https://api.dev.nva.aws.unit.no/publication/018a3c252deb-7fa96a4f-aa10-4fa8-aba6-209bdca078bf"
-      },
-      "data": {
-        "m": {
-          "internationalCollaboration": {
-            "bool": false
-          },
-          "level": {
-            "s": "1"
-          },
-          "creators": {
-            "l": [
-              {
-                "m": {
-                  "creatorId": {
-                    "s": "https://api.dev.nva.aws.unit.no/cristin/person/1136326"
-                  },
-                  "affiliations": {
-                    "l": [
-                      {
-                        "s": "https://api.dev.nva.aws.unit.no/cristin/organization/20754.3.0.0"
-                      }
-                    ]
-                  }
-                }
-              }
-            ]
-          },
-          "instanceType": {
-            "s": "AcademicArticle"
-          },
-          "applicable": {
-            "bool": true
-          },
-          "creatorCount": {
-            "n": "0"
-          },
-          "publicationId": {
-            "s": "https://api.dev.nva.aws.unit.no/publication/018a3c252deb-7fa96a4f-aa10-4fa8-aba6-209bdca078bf"
-          },
-          "publicationDate": {
-            "m": {
-              "month": {
-                "s": "8"
-              },
-              "year": {
-                "s": "2023"
-              },
-              "day": {
-                "s": "28"
-              }
-            }
-          },
-          "approvalStatuses": {
-            "l": [
-              {
-                "m": {
-                  "institutionId": {
-                    "s": "https://api.dev.nva.aws.unit.no/cristin/organization/20754.0.0.0"
-                  },
-                  "status": {
-                    "s": "Pending"
-                  }
-                }
-              }
-            ]
-          }
-        }
-      },
-      "SecondaryIndex1RangeKey": {
-        "s": "https://api.dev.nva.aws.unit.no/publication/018a3c252deb-7fa96a4f-aa10-4fa8-aba6-209bdca078bf"
-      },
-      "PrimaryKeyHashKey": {
-        "s": "CANDIDATE#167ee953-b8a9-4ef6-a3c7-a62519ed6a6e"
-      },
-      "PrimaryKeyRangeKey": {
-        "s": "CANDIDATE#167ee953-b8a9-4ef6-a3c7-a62519ed6a6e"
-      }
-    },
-    "sequenceNumber": "3750100000000011651060377",
-    "sizeBytes": 1056,
-    "streamViewType": "NEW_AND_OLD_IMAGES"
-  },
-  "eventSourceARN": "arn:aws:dynamodb:eu-west-1:884807050265:table/nva-nvi-master-pipelines-NvaNvi-1LZJ4B13FNYPO-nva-nvi/stream/2023-08-29T09:19:20.077"
-}
