--- index-handlers/src/test/resources/genericDynamoDbRecord.json
@@ -1,18 +1,0 @@
-{
-  "dynamodb": {
-    "keys": {
-      "PrimaryKeyRangeKey": {
-        "s": "CANDIDATE#__REPLACE_IDENTIFIER__"
-      },
-      "PrimaryKeyHashKey": {
-        "s": "CANDIDATE#__REPLACE_IDENTIFIER__"
-      }
-    },
-    "newImage": {
-      "identifier": {
-        "s": "__REPLACE_IDENTIFIER__"
-      }
-    }
-  },
-  "eventSourceARN": "arn:aws:dynamodb:eu-west-1:884807050265:table/nva-nvi-master-pipelines-NvaNvi-1LZJ4B13FNYPO-nva-nvi/stream/2023-08-29T09:19:20.077"
-}
